# swagger_server/swagger/swagger.yaml
openapi: 3.0.1
info:
  title: Wypożyczalnia Aut - API
  description: API do obsługi wypożyczalni
  version: "1.0"
servers:
  - url: http://localhost:5000/
...
paths:
...
  /admin/car:
    get:
      tags:
        - admin
      summary: Pobiera liste aut
      operationId: admin_car_get
      parameters:
        - name: Session_Token
          in: header
          description: Token sesji
          required: true
          style: simple
          explode: false
          schema:
            type: integer
            example: 12943854693802
          example: "1111111111111111"
        - name: locationLat
          in: query
          description: Szerokosc geograficzna
          required: false
          style: form
          explode: true
          schema:
            type: string
          example: "51.312312"
        - name: locationLong
          in: query
          description: Dlugosc geograficzna
          required: false
          style: form
          explode: true
          schema:
            type: string
          example: "16.312312"
        - name: pagelength
          in: query
          description: Number of records to return
          required: false
          style: form
          explode: true
          schema:
            maximum: 20
            minimum: 1
            type: number
            default: 5
        - name: startindex
          in: query
          description: Start index for paging
          required: false
          style: form
          explode: true
          schema:
            minimum: 0
            type: number
            default: 0
        - name: distance
          in: query
          description: "Dystans w m, w jakim szukac. Wartosci wieksze od 2000m beda\
          \ automatycznie zamieniane na 1km, mniejsze od 100m na 100m"
          required: false
          style: form
          explode: true
          schema:
            maximum: 2000
            minimum: 100
            type: integer
        - name: details
          in: query
          description: Szczegoly auta
          required: false
          style: form
          explode: true
          schema:
            type: boolean
            default: false
      responses:
        "200":
          description: Lista aut
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/inline_response_200_5'
        "400":
          description: "Ktorys argument jest niepoprawny. To jest api admina, nie\
            \ ma znaczenia Ktory"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Empty'
        "401":
          description: Brak autoryzacji! Sprawdz token!
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Empty'
      x-openapi-router-controller: swagger_server.controllers.admin_controller
    post:
      tags:
        - admin
      summary: Dodaje nowe auto
      operationId: admin_car_post
      parameters:
        - name: Session_Token
          in: header
          description: Token sesji
          required: true
          style: simple
          explode: false
          schema:
            type: integer
            example: 12943854693802
          example: "1111111111111111"
      requestBody:
        description: "Ignorujcie parametry \"bez sensu\" np last used, last rental\
          \ itp, nie chce mi sie tworzyc specjalnie scheme na to XD"
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/admin_car_body'
        required: true
      responses:
        "200":
          description: Successfully added new car!
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/inline_response_200_6'
        "400":
          description: Car not added! Check arguments!
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Empty'
        "401":
          description: Brak autoryzacji! Sprawdz token!
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Empty'
      x-openapi-router-controller: swagger_server.controllers.admin_controller
  /admin/car/{id}:
    get:
      tags:
        - admin
      summary: Pobiera szczegoly auta
      operationId: admin_car_id_get
      parameters:
        - name: Session_Token
          in: header
          description: Token sesji
          required: true
          style: simple
          explode: false
          schema:
            type: integer
            example: 12943854693802
          example: "1111111111111111"
        - name: id
          in: path
          description: Id auta
          required: true
          style: simple
          explode: false
          schema:
            type: integer
      responses:
        "200":
          description: Auto
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/inline_response_200_7'
        "400":
          description: niewlasiwe ID auta
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Empty'
        "401":
          description: Brak autoryzacji! Sprawdz token!
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Empty'
      x-openapi-router-controller: swagger_server.controllers.admin_controller
    delete:
      tags:
        - admin
      summary: Usuwa auto z systemu
      operationId: admin_car_id_delete
      parameters:
        - name: Session_Token
          in: header
          description: Token sesji
          required: true
          style: simple
          explode: false
          schema:
            type: integer
            example: 12943854693802
          example: "1111111111111111"
        - name: id
          in: path
          description: Id auta
          required: true
          style: simple
          explode: false
          schema:
            type: integer
      responses:
        "200":
          description: Usunieto auto!
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Empty'
        "400":
          description: Auto nie moze byc usuniete z jakiegos powodu!
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Empty'
        "401":
          description: Brak autoryzacji! Sprawdz token!
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Empty'
      x-openapi-router-controller: swagger_server.controllers.admin_controller
    patch:
      tags:
        - admin
      summary: Modyfikuje dane auta
      operationId: admin_car_id_patch
      parameters:
        - name: Session_Token
          in: header
          description: Token sesji
          required: true
          style: simple
          explode: false
          schema:
            type: integer
            example: 12943854693802
          example: "1111111111111111"
        - name: id
          in: path
          description: Id auta
          required: true
          style: simple
          explode: false
          schema:
            type: integer
      requestBody:
        description: "Co zmienic? Dawac w Jsonie tylko te wartosci, ktore chcesz zmienic\
          \ (oczywiskie niektorych rzeczy sie nie da, np modyfikowanie ID)"
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/car_id_body'
      responses:
        "200":
          description: Zmodyfikowano!
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Empty'
        "400":
          description: Nie udalo sie zmodyfikowac
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Empty'
        "401":
          description: Brak autoryzacji! Sprawdz token!
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Empty'
      x-openapi-router-controller: swagger_server.controllers.admin_controller
  /admin/car/{id}/rentalhistory:
    get:
      tags:
        - admin
      summary: Zwraca historie przejazdow
      operationId: admin_car_id_rentalhistory_get
      parameters:
        - name: id
          in: path
          description: Id auta
          required: true
          style: simple
          explode: false
          schema:
            type: integer
        - name: Session_Token
          in: header
          description: Token sesji
          required: true
          style: simple
          explode: false
          schema:
            type: integer
            example: 12943854693802
          example: "1111111111111111"
        - name: startindex
          in: query
          description: Start index for paging
          required: false
          style: form
          explode: true
          schema:
            minimum: 0
            type: number
            default: 0
        - name: pagelength
          in: query
          description: Number of records to return
          required: false
          style: form
          explode: true
          schema:
            maximum: 20
            minimum: 1
            type: number
            default: 5
      responses:
        "200":
          description: Historia przejazdow
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/inline_response_200_8'
        "400":
          description: Blad argumentow
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Empty'
        "401":
          description: Brak autoryzacji! Sprawdz token!
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Empty'
      x-openapi-router-controller: swagger_server.controllers.admin_controller
  /admin/car/{id}/reservationhistory:
    get:
      tags:
        - admin
      summary: Zwraca historie rezerwacji
      operationId: admin_car_id_reservationhistory_get
      parameters:
        - name: id
          in: path
          description: Id auta
          required: true
          style: simple
          explode: false
          schema:
            type: integer
        - name: Session_Token
          in: header
          description: Token sesji
          required: true
          style: simple
          explode: false
          schema:
            type: integer
            example: 12943854693802
          example: "1111111111111111"
        - name: startindex
          in: query
          description: Start index for paging
          required: false
          style: form
          explode: true
          schema:
            minimum: 0
            type: number
            default: 0
        - name: pagelength
          in: query
          description: Number of records to return
          required: false
          style: form
          explode: true
          schema:
            maximum: 20
            minimum: 1
            type: number
            default: 5
      responses:
        "200":
          description: Historia rezerwacji
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/inline_response_200_9'
        "400":
          description: Blad argumentow
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Empty'
        "401":
          description: Brak autoryzacji! Sprawdz token!
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Empty'
      x-openapi-router-controller: swagger_server.controllers.admin_controller
  /admin/user:
    get:
      tags:
        - admin
      summary: Zwraca liste uzytkownikow
      operationId: admin_user_get
      parameters:
        - name: Session_Token
          in: header
          description: Token sesji
          required: true
          style: simple
          explode: false
          schema:
            type: integer
            example: 12943854693802
          example: "1111111111111111"
        - name: startindex
          in: query
          description: Start index for paging
          required: false
          style: form
          explode: true
          schema:
            minimum: 0
            type: number
            default: 0
        - name: pagelength
          in: query
          description: Number of records to return
          required: false
          style: form
          explode: true
          schema:
            maximum: 20
            minimum: 1
            type: number
            default: 5
        - name: details
          in: query
          description: Szczegoly auta
          required: false
          style: form
          explode: true
          schema:
            type: boolean
            default: false
        - name: filter
          in: query
          description: Login lub imie lub nazwisko lub email usera
          required: false
          style: form
          explode: true
          schema:
            type: string
            default: ""
      responses:
        "200":
          description: Lista uzytkownikow
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/inline_response_200_10'
        "400":
          description: "Ktorys argument jest niepoprawny. To jest api admina, nie\
            \ ma znaczenia Ktory"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Empty'
        "401":
          description: Brak autoryzacji! Sprawdz token!
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Empty'
      x-openapi-router-controller: swagger_server.controllers.admin_controller
  /admin/user/{id}:
    get:
      tags:
        - admin
      summary: Pobiera szczegoly uzytkownika
      operationId: admin_user_id_get
      parameters:
        - name: Session_Token
          in: header
          description: Token sesji
          required: true
          style: simple
          explode: false
          schema:
            type: integer
            example: 12943854693802
          example: "1111111111111111"
        - name: id
          in: path
          description: Id usera
          required: true
          style: simple
          explode: false
          schema:
            type: integer
      responses:
        "200":
          description: User
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/inline_response_200_11'
        "400":
          description: niewlasiwe ID usera
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Empty'
        "401":
          description: Brak autoryzacji! Sprawdz token!
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Empty'
      x-openapi-router-controller: swagger_server.controllers.admin_controller
    delete:
      tags:
        - admin
      summary: Usuwa usera z systemu
      operationId: admin_user_id_delete
      parameters:
        - name: Session_Token
          in: header
          description: Token sesji
          required: true
          style: simple
          explode: false
          schema:
            type: integer
            example: 12943854693802
          example: "1111111111111111"
        - name: id
          in: path
          description: Id usera
          required: true
          style: simple
          explode: false
          schema:
            type: integer
      responses:
        "200":
          description: Usunieto usera!
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Empty'
        "400":
          description: User nie moze byc usuniete z jakiegos powodu!
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Empty'
        "401":
          description: Brak autoryzacji! Sprawdz token!
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Empty'
      x-openapi-router-controller: swagger_server.controllers.admin_controller
    patch:
      tags:
        - admin
      summary: Modyfikuje dane usera
      operationId: admin_user_id_patch
      parameters:
        - name: Session_Token
          in: header
          description: Token sesji
          required: true
          style: simple
          explode: false
          schema:
            type: integer
            example: 12943854693802
          example: "1111111111111111"
        - name: id
          in: path
          description: Id usera
          required: true
          style: simple
          explode: false
          schema:
            type: integer
      requestBody:
        description: "Co zmienic? Dawac w Jsonie tylko te wartosci, ktore chcesz zmienic\
          \ (oczywiskie niektorych rzeczy sie nie da, np modyfikowanie ID). Nie modyfikowac\
          \ np aktywacji lub licencji, od tego jest inne api"
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/user_id_body'
      responses:
        "200":
          description: Zmodyfikowano!
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Empty'
        "400":
          description: Nie udalo sie zmodyfikowac
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Empty'
        "401":
          description: Brak autoryzacji! Sprawdz token!
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Empty'
      x-openapi-router-controller: swagger_server.controllers.admin_controller
  /admin/user/{id}/activate:
    post:
      tags:
        - admin
      summary: Aktywuje konto usera
      operationId: admin_user_id_activate_post
      parameters:
        - name: Session_Token
          in: header
          description: Token sesji
          required: true
          style: simple
          explode: false
          schema:
            type: integer
            example: 12943854693802
          example: "1111111111111111"
        - name: id
          in: path
          description: Id usera
          required: true
          style: simple
          explode: false
          schema:
            type: integer
      responses:
        "200":
          description: Aktywowano!
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Empty'
        "400":
          description: Nie udalo sie Aktywowac/konto juz aktywne
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Empty'
        "401":
          description: Brak autoryzacji! Sprawdz token!
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Empty'
      x-openapi-router-controller: swagger_server.controllers.admin_controller
...
  /admin/user/{id}/rentalhistory:
    get:
      tags:
        - admin
      summary: Zwraca historie przejazdow
      operationId: admin_user_id_rentalhistory_get
      parameters:
        - name: id
          in: path
          description: Id usera
          required: true
          style: simple
          explode: false
          schema:
            type: integer
        - name: Session_Token
          in: header
          description: Token sesji
          required: true
          style: simple
          explode: false
          schema:
            type: integer
            example: 12943854693802
          example: "1111111111111111"
        - name: startindex
          in: query
          description: Start index for paging
          required: false
          style: form
          explode: true
          schema:
            minimum: 0
            type: number
            default: 0
        - name: pagelength
          in: query
          description: Number of records to return
          required: false
          style: form
          explode: true
          schema:
            maximum: 20
            minimum: 1
            type: number
            default: 5
      responses:
        "200":
          description: Historia przejazdow
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/inline_response_200_8'
        "400":
          description: Blad argumentow
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Empty'
        "401":
          description: Brak autoryzacji! Sprawdz token!
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Empty'
      x-openapi-router-controller: swagger_server.controllers.admin_controller
...
components:
  schemas:
    Empty:
      type: object
      properties: { }
      description: Pusty Json
      example: { }
...
    inline_response_200_5:
      type: object
      properties:
        cars:
          type: array
          items:
            $ref: '#/components/schemas/inline_response_200_7'
      example:
        cars:
          - car:
              locationLat: "51.235123"
              charge: 100
              distanceLeft: 20
              seats: 5
              carId: 1248530713
              regNumber: DW112233
              locationLong: "16.50312"
              kmCost: "1.30"
              model: Yaris 1.0
              activationCost: "10.30"
              timeCost: "0.30"
              brand: Toyota
              status: ACTIVE
            details:
              lastUsed: 16123234234
              eSimImei: "125123123"
              esimPhoneNumber: 125123123
              lastReservation: 1238912093
              lastUpdate: 125123123
              vin: ASDSGFA123123123
              lastRental: 0
              lastService: 125123123
              mileage: 125123123
          - car:
              locationLat: "51.235123"
              charge: 100
              distanceLeft: 20
              seats: 5
              carId: 1248530713
              regNumber: DW112233
              locationLong: "16.50312"
              kmCost: "1.30"
              model: Yaris 1.0
              activationCost: "10.30"
              timeCost: "0.30"
              brand: Toyota
              status: ACTIVE
            details:
              lastUsed: 16123234234
              eSimImei: "125123123"
              esimPhoneNumber: 125123123
              lastReservation: 1238912093
              lastUpdate: 125123123
              vin: ASDSGFA123123123
              lastRental: 0
              lastService: 125123123
              mileage: 125123123
    admin_car_body:
      type: object
      allOf:
        - $ref: '#/components/schemas/Car'
        - $ref: '#/components/schemas/CarDetails'
    inline_response_200_6:
      properties:
        carId:
          type: integer
          description: Id of a new added car
          example: 12319828391283
      example:
        carId: 12319828391283
...
    car_id_body:
      type: object
      properties:
        patch:
          type: object
          description: "Co zmienic? Tylko te wartosci, ktore chcesz zmienic (oczywiskie\
            \ niektorych rzeczy sie nie da, np modyfikowanie ID)"
          allOf:
            - $ref: '#/components/schemas/Car'
            - $ref: '#/components/schemas/CarDetails'
...
    user_id_body:
      type: object
      properties:
        patch:
          type: object
          description: "Co zmienic? Tylko te wartosci, ktore chcesz zmienic (oczywiskie\
            \ niektorych rzeczy sie nie da, np modyfikowanie ID)"
          allOf:
            - $ref: '#/components/schemas/User'
            - $ref: '#/components/schemas/UserDetails'
...